redux 도입중 + 리팩토링
