+++ src/tools/CommonFunc.ts
@@ -8,3 +8,30 @@ export const changeStringToJSON = (data: any) => {
     }
   }
 };
+
+export function calculateDistance(
+  lat1: number,
+  lon1: number,
+  lat2: number,
+  lon2: number,
+): number {
+  const earthRadius = 6371; // Radius of the Earth in kilometers
+  const dLat = toRadians(lat2 - lat1);
+  const dLon = toRadians(lon2 - lon1);
+
+  const a =
+    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+    Math.cos(toRadians(lat1)) *
+      Math.cos(toRadians(lat2)) *
+      Math.sin(dLon / 2) *
+      Math.sin(dLon / 2);
+
+  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+
+  const distance = earthRadius * c * 1000; // Distance in meters
+  return Math.round(distance);
+}
+
+export function toRadians(degrees: number): number {
+  return degrees * (Math.PI / 180);
+}
